+++ TKMovesets2/GameInteractions/Extractors/Extractor.cpp
@@ -55,13 +55,13 @@ namespace ExtractorUtils
 
 		gameAddr animPtr = anim_postBoneDescriptorAddr + vv73 + aa4;
 
-		int baseFrame = animLength - 1 - 1;
+		int baseFrame = (int)(animLength - 1 - 1);
 		int keyframe = baseFrame / 16;
 		unsigned int _v56_intPtr = (unsigned int)process->readInt32(animPtr + 4 * keyframe);
 		_v56_intPtr = SWAP_INT32(_v56_intPtr);
 
 		gameAddr animPtr_2 = animPtr + _v56_intPtr;
-		int lastArg_copy = boneCount;
+		int lastArg_copy = (int)boneCount;
 
 		do
 		{
@@ -98,7 +98,7 @@ namespace ExtractorUtils
 		unsigned int _v56_intPtr = (unsigned int)process->readInt32(animPtr + 4 * keyframe);
 
 		gameAddr animPtr_2 = animPtr + _v56_intPtr;
-		int lastArg_copy = boneCount;
+		int lastArg_copy = (int)boneCount;
 
 		do
 		{
@@ -116,7 +116,7 @@ namespace ExtractorUtils
 	void WriteFileData(std::ofstream &file, std::vector<std::pair<Byte*, uint64_t>>& blocks, uint8_t&progress, uint8_t progress_max)
 	{
 		uint8_t remainingProgress = progress_max - progress;
-		uint8_t step = remainingProgress / blocks.size();
+		uint8_t step = (size_t)remainingProgress / blocks.size();
 
 		for (const std::pair<Byte*, uint64_t>& block : blocks) {
 			char* blockData = (char*)block.first;
